[REFACTOR/#92] 뷰모델 STATE 관리 변경
