+++ feature/src/main/java/com/terning/feature/search/search/SearchViewModel.kt
@@ -5,6 +5,8 @@ import androidx.lifecycle.viewModelScope
 import com.terning.core.state.UiState
 import com.terning.domain.repository.SearchRepository
 import com.terning.feature.R
+import com.terning.feature.search.search.model.SearchScrapsListState
+import com.terning.feature.search.search.model.SearchViewsListState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -18,11 +20,13 @@ import javax.inject.Inject
 class SearchViewModel @Inject constructor(
     private val searchRepository: SearchRepository,
 ) : ViewModel() {
-    private val _viewState: MutableStateFlow<SearchState> = MutableStateFlow(SearchState())
-    val viewState: StateFlow<SearchState> = _viewState.asStateFlow()
+    private val _viewState: MutableStateFlow<SearchViewsListState> =
+        MutableStateFlow(SearchViewsListState())
+    val viewState: StateFlow<SearchViewsListState> = _viewState.asStateFlow()
 
-    private val _scrapState: MutableStateFlow<SearchState> = MutableStateFlow(SearchState())
-    val scrapState: StateFlow<SearchState> = _scrapState.asStateFlow()
+    private val _scrapState: MutableStateFlow<SearchScrapsListState> =
+        MutableStateFlow(SearchScrapsListState())
+    val scrapState: StateFlow<SearchScrapsListState> = _scrapState.asStateFlow()
 
     private val _sideEffect: MutableSharedFlow<SearchSideEffect> = MutableSharedFlow()
     val sideEffect = _sideEffect.asSharedFlow()
