+++ feature/src/main/java/com/terning/feature/search/search/model/SearchScrapsListState.kt
@@ -1,9 +1,8 @@
-package com.terning.feature.search.search
+package com.terning.feature.search.search.model
 
 import InternshipAnnouncementModel
 import com.terning.core.state.UiState
 
-data class SearchState(
-    var searchViewsList: UiState<List<InternshipAnnouncementModel>> = UiState.Loading,
+data class SearchScrapsListState(
     var searchScrapsList: UiState<List<InternshipAnnouncementModel>> = UiState.Loading,
 )
