+++ feature/src/main/java/com/terning/feature/search/search/model/SearchViewsListState.kt
@@ -1,0 +1,8 @@
+package com.terning.feature.search.search.model
+
+import InternshipAnnouncementModel
+import com.terning.core.state.UiState
+
+data class SearchViewsListState(
+    var searchViewsList: UiState<List<InternshipAnnouncementModel>> = UiState.Loading,
+)
